requirements.txt
streamlit>=1.44
openai>=1.30.0
requests>=2.31.0
tencentcloud-sdk-python>=3.0.0
